# ghoststake-backend/openapi.yaml
openapi: 3.1.0
info:
  title: GhostStake API
  version: "1.0.0"
  summary: Read the protocol's derived state without talking to a chain.
  description: |
    Every figure here is derived from public chain state — either from the
    append-only ledger the indexer writes, or from a live `eth_call` pinned to
    one block. Nothing is authoritative: the contracts are. This exists so a
    client does not need a multicall and an ABI to render a page.

    ## Two conventions that will otherwise catch you out

    **Every uint256 is a decimal string, never a JSON number.** JSON numbers
    are IEEE-754 doubles and a token balance in wei exceeds their 53 bits of
    integer precision routinely, so `JSON.parse` would silently fabricate the
    low digits. Feed these to `BigInt()`.

    **`healthFactor` and `ltv` are `null` when there is no debt.** The contract
    returns `type(uint256).max`, which is correct on-chain and a 78-digit
    integer in JSON that every consumer would have to special-case. `hasDebt`
    tells you which case you are in.

    ## Freshness

    Indexed reads are deliberately behind the chain head: the indexer stays
    `INDEXER_CONFIRMATIONS` blocks back, because a shallower block can still be
    reorged out. Every indexed response carries `indexedBlock` and `asOf` —
    show them rather than implying the data is live. A user who has just staked
    and does not see it yet has found the confirmation lag, not a bug.

...
servers:
...
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/nonce:
    post:
      tags: [auth]
      summary: Request a login challenge
      description: |
        Returns the exact SIWE message to sign. The server stores what it
        rendered and verifies against its own copy, so the client must sign
        this string verbatim and never compose or edit one.

        Rate limited to 20 requests per minute per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address:
                  $ref: "#/components/schemas/Address"
      responses:
        "200":
          description: Challenge issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  nonce: { type: string }
                  message:
                    type: string
                    description: Sign this exactly as given.
                  expiresAt: { type: string, format: date-time }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/verify:
    post:
      tags: [auth]
      summary: Exchange a signature for a token
      description: |
        The address is resolved server-side from the nonce, so a client cannot
        nominate one. Failures are deliberately vague: a specific reason would
        make this an oracle for probing valid addresses and live nonces.

        Rate limited to 20 requests per minute per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nonce, signature]
              properties:
                nonce: { type: string }
                signature: { type: string }
      responses:
        "200":
          description: Signature verified.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: JWT bearer token. }
                  address: { $ref: "#/components/schemas/Address" }
                  expiresAt: { type: string, format: date-time }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/me:
    get:
      tags: [auth]
      summary: The authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  address: { $ref: "#/components/schemas/Address" }
                  createdAt: { type: string, format: date-time }
                  lastLoginAt: { type: [string, "null"], format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/rounds:
    get:
      tags: [market]
      summary: Recent rounds with pool splits
      description: |
        Newest first. Pools are summed from indexed `PositionTaken` events, so
        a client never has to sum on-chain.

        Unauthenticated on purpose: every figure is derived from public chain
        state, and requiring a login to read a public blockchain would be
        theatre. Rate limited to 120 requests per minute per IP.
      security: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/MarketQuery"
      responses:
        "200":
          description: Rounds, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IndexedEnvelope"
                  - type: object
                    properties:
                      rounds:
                        type: array
                        items: { $ref: "#/components/schemas/Round" }
                      degraded:
                        $ref: "#/components/schemas/DegradedMarkets"
  /api/v1/positions/{address}:
    get:
      tags: [market]
      summary: An address's positions, open and historical
      description: |
        `open` is positions in rounds still running; `history` is everything
        settled or voided. Each entry carries the whole round, so a client does
        not have to join two responses.

        Rate limited to 120 requests per minute per IP.
      security: []
      parameters:
        - $ref: "#/components/parameters/AddressPath"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/MarketQuery"
      responses:
        "200":
          description: Positions.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IndexedEnvelope"
                  - type: object
                    properties:
                      address: { $ref: "#/components/schemas/Address" }
                      open:
                        type: array
                        items: { $ref: "#/components/schemas/Position" }
                      history:
                        type: array
                        items: { $ref: "#/components/schemas/Position" }
                      degraded:
                        $ref: "#/components/schemas/DegradedMarkets"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/positions/at-risk:
    get:
      tags: [market]
      summary: Borrowers close to liquidation, sickest first
      description: |
        The discovery step liquidation was missing. `liquidate` is
        permissionless, which is what keeps the protocol solvent — but every
        other view is per-address, so a liquidator had to already know whose
        position was underwater before they could act on it.

        Two sources, each asked what only it can answer. The **ledger** supplies
        the names: it holds every borrow and repayment, and the chain has no
        borrower enumeration at all. The **chain** supplies the numbers: a health
        factor is scaled debt times an index that advances every second and
        emits nothing, so there is no event to index for it. Every figure here
        is read at `block`, which is the chain head, not at `indexedBlock`.

        Ordered by `healthFactor` ascending. Anything below `1e18` is
        liquidatable now; above it is a watchlist.

        `profitable` is the field to act on. Past the point where collateral
        covers debt plus the liquidation bonus, no repayment size leaves a
        liquidator ahead — those rows carry `writeOffCandidate: true` and are
        closed with `writeOffBadDebt`, not with `liquidate`.

        Rate limited to 120 requests per minute per IP.
      security: []
      parameters:
        - name: limit
          in: query
          required: false
          description: |
            How many borrowers to read, default 50, maximum 200. A bound on
            `eth_call`s rather than on rows: each borrower costs four chain
            reads. When the cap bites, the borrowers scanned are those with the
            largest debt — a truncated list should be missing the trivia rather
            than the risk — and `truncated` says so.
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Borrowers carrying debt, ordered by health factor.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IndexedEnvelope"
                  - type: object
                    properties:
                      block:
                        type: integer
                        description: |
                          The chain block every figure below was read at. One
                          block for all of them: this is a ranking, and health
                          factors sampled at different blocks are ordered by
                          nothing in particular.
                      scanned: { type: integer }
                      truncated:
                        type: boolean
                        description: The cap was reached; there may be more.
                      positions:
                        type: array
                        items: { $ref: "#/components/schemas/AtRiskPosition" }
        "503":
          description: The service was started without a chain connection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/activity/{address}:
    get:
      tags: [market]
      summary: Everything one address has done, newest first
      description: |
        One merged, paginated feed over both halves of the protocol: the
        lending ledger and the round history. They are merged here rather than
        client-side because the two interleave — a user borrows and then stakes
        what they borrowed, and those rows belong next to each other — and
        because merging two independently-limited lists gets the page boundary
        wrong in a way nothing surfaces.

        Every amount is nominal, as emitted at the time. Index-scaled balance
        entries are excluded at the query: a scaled amount is what makes a
        balance summable and exactly what makes it wrong as history, because
        rendering it means either showing a number the user never saw or
        converting it with today's index — a completed transaction whose
        figure changes on every reload.

        Paged on the log's own coordinates rather than an offset. The indexer
        writes to the head continuously, so an offset page shifts under a
        reader mid-scroll. Pass `nextCursor` back as `?cursor=`; `null` means
        this was the last page.

        Rate limited to 120 requests per minute per IP.
      security: []
      parameters:
        - $ref: "#/components/parameters/AddressPath"
        - $ref: "#/components/parameters/ActivityLimit"
        - $ref: "#/components/parameters/ActivityCursor"
      responses:
        "200":
          description: Activity.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IndexedEnvelope"
                  - type: object
                    properties:
                      address: { $ref: "#/components/schemas/Address" }
                      events:
                        type: array
                        items: { $ref: "#/components/schemas/ActivityEvent" }
                      nextCursor:
                        type: string
                        nullable: true
                        description: |
                          Pass back as `?cursor=` for the next page. Null on
                          the last page.
                        example: "9421133-4-0"
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/ws:
    get:
      tags: [market]
      summary: Live round updates (WebSocket)
      description: |
        A WebSocket upgrade, not a JSON endpoint — OpenAPI cannot describe the
        frames, so they are documented here.

        Connect to `wss://<host>/api/v1/ws`, optionally with `?address=0x…` to
        also receive that address's positions. The origin is checked against
        the same allowlist as CORS; a browser cannot be tricked into opening
        this from another site.

        The server sends one frame on connect and another whenever an indexed
        range touches a round (or, if you named an address, that account).
        Every frame is a **complete snapshot**, never a delta — a client that
        misses one is not left inconsistent, and a slow client is dropped
        rather than buffered without bound.

        Frame shape:

        ```json
        {
          "type": "rounds",
          "block": 11562445,
          "chainId": 11155111,
          "asOf": "2026-08-25T06:59:58Z",
          "rounds": [ /* Round */ ],
          "address": "0x…",
          "positions": [ /* Position, only when ?address= was given */ ]
        }
        ```

        The client is not expected to send anything. Frames it does send are
        read and discarded, which is what keeps the read deadline and the
        pong handler alive.
      security: []
      parameters:
        - name: address
          in: query
          required: false
          description: Also stream this address's positions.
          schema: { $ref: "#/components/schemas/Address" }
      responses:
        "101":
          description: Switching protocols; the WebSocket is open.
        "400":
          description: Malformed address, or not a valid WebSocket upgrade.
        "403":
          description: Origin not in the allowlist.
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AddressPath:
      name: address
      in: path
      required: true
      description: Ethereum address. Case-insensitive; normalised to EIP-55.
      schema: { $ref: "#/components/schemas/Address" }
    MarketQuery:
      name: market
      in: query
      required: false
      description: |
        Narrow to one market, by its ParimutuelRound address. Omit for every
        indexed market.

        Rounds from different markets are unrelated: ids restart at 1 in each
        one, so `id` is not an identity across a mixed listing and `market`
        is the other half of it. A malformed address is a 400 rather than
        being ignored — ignoring it would widen the answer rather than narrow
        it, and the caller has no way to tell.
      schema: { $ref: "#/components/schemas/Address" }
    ActivityLimit:
      name: limit
      in: query
      required: false
      description: |
        Activity rows to return. Defaults to 50, clamped to 200. Higher than
        the round limit because a row here is one already-decoded database row
        rather than a whole round projected and priced. Unparseable or
        non-positive values fall back to the default rather than erroring.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    ActivityCursor:
      name: cursor
      in: query
      required: false
      description: |
        A position in the feed, as `<blockNumber>-<logIndex>-<recordIndex>`.
        Take it from the previous response's `nextCursor` and pass it
        unchanged; omit for the first page.

        The log's coordinates rather than an offset, because the indexer
        writes to the head while a client is reading: an offset page shifts
        under it and repeats or drops rows at the boundary. A malformed cursor
        is a 400, not a silent first page — answering a request for page four
        with page one reads as a list that never ends.
      schema:
        type: string
        example: "9421133-4-0"
    Limit:
      name: limit
      in: query
      required: false
      description: |
        Rounds to return. Defaults to 20, clamped to 100 — the query behind it
        reads every event of every round it returns, so an unbounded limit is a
        request that reads the whole table. Values that are unparseable or
        non-positive fall back to the default rather than erroring.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid or expired token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such record.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests from this IP.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    DegradedMarkets:
      type: array
      description: |
        Markets that could not be described, so a short listing says so rather
        than quietly being a short listing.

        A market's rake, entry cutoff and minimum side pool are read from its
        contract. Round history deliberately spans deployments, so a listing
        can contain markets this process has never watched — and before GHO-82
        one of them being unreadable failed the whole request, including the
        settled history that needs no live read at all.

        Rounds and positions belonging to a listed market are omitted from this
        response. They are not gone; they could not be priced. Absent when
        every market was readable, which is the ordinary case.
      items: { $ref: "#/components/schemas/Address" }
    Error:
      type: object
      properties:
        error:
          type: string
          description: |
            Deliberately generic on auth failures — a specific reason would
            make the response an oracle for probing addresses and nonces.
    Address:
      type: string
      pattern: "^0x[0-9a-fA-F]{40}$"
      examples: ["0xE64b1081681c475A0Ec61b3B3f90e82734049275"]
    Uint256:
      type: string
      pattern: "^[0-9]+$"
      description: |
        A uint256 as a decimal string. Never a JSON number — see the note on
        precision in the API description. Parse with `BigInt()`.
      examples: ["5000000000"]
    IndexedEnvelope:
      type: object
      description: Fields every indexed response carries, describing its freshness.
      properties:
        chainId: { type: integer, examples: [11155111] }
        indexedBlock:
          type: integer
          description: |
            Highest block the indexer has processed. Behind the chain head by
            the configured confirmation depth.
        asOf:
          type: string
          format: date-time
          description: When the response was rendered. Always UTC.
    AtRiskPosition:
      type: object
      description: |
        One borrower, and what closing their position would cost and pay. Every
        figure mirrors `CollateralVault.liquidate` step for step.
      properties:
        address: { $ref: "#/components/schemas/Address" }
        collateral: { $ref: "#/components/schemas/Uint256" }
        debt:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: |
            Includes interest pending since the pool last accrued, because a
            liquidator's transaction accrues before it reads the health factor.
            This is the debt they would actually find.
        healthFactor:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: WAD. Below 1e18 is liquidatable.
        ltv: { $ref: "#/components/schemas/Uint256" }
        band:
          type: string
          enum: [none, safe, caution, danger]
          description: The same bands the dashboard uses, so "at risk" means one thing.
        liquidatable: { type: boolean }
        maxRepay:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: |
            The most one call may clear: the close factor applied to the debt,
            lifted to the whole lien once the position is far enough under.
        seized:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: |
            Collateral that repayment takes, bonus included and capped at what
            the position holds. The cap is why `profitable` can be false.
        bonus:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: Seized less repaid. Zero, never negative — see `profitable`.
        profitable:
          type: boolean
          description: |
            Whether a rational liquidator would call. False is not an edge
            case: past the point where collateral covers debt plus bonus, the
            caller is out of pocket, which is why these positions sit.
        fullLiquidation:
          type: boolean
          description: The close factor has lifted, so one call clears the whole lien.
        writeOffCandidate:
          type: boolean
          description: |
            Owes more than it holds. No liquidation can close it; `writeOffBadDebt`
            can, and pays nothing for doing so.
    Round:
      type: object
      properties:
        market:
          allOf: [{ $ref: "#/components/schemas/Address" }]
          description: |
            The ParimutuelRound this round belongs to.

            Round ids restart at 1 in every market, so `id` alone does not
            identify a round — a client keying a list on it merges two
            markets' rounds into one row. The pair `(market, id)` is the
            identity.
        id: { type: integer }
        status:
          type: string
          enum: [none, open, locked, resolved, void]
          description: Stored state on the contract.
        phase:
          type: string
          enum: [none, open, cutoff, observation, resolved, void]
          description: |
            What the round looks like from outside, including the two phases
            that are functions of the clock rather than of storage. `cutoff`
            and `observation` never appear as `status`.
        entryOpen:
          type: boolean
          description: |
            Whether a stake would be accepted *now*. False during `cutoff`
            even though the round has not locked — which is why a UI must read
            this rather than inferring it from `phase == open`.
        openTime: { type: string, format: date-time }
        lockTime: { type: string, format: date-time }
        closeTime: { type: string, format: date-time }
        upPool: { $ref: "#/components/schemas/Uint256" }
        downPool: { $ref: "#/components/schemas/Uint256" }
        totalPool: { $ref: "#/components/schemas/Uint256" }
        upOdds:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: |
            WAD-scaled payout multiplier for a unit staked on this side, net of
            rake. 1e18 is break-even.
        downOdds: { $ref: "#/components/schemas/Uint256" }
        lockPrice:
          oneOf:
            - $ref: "#/components/schemas/Uint256"
            - type: "null"
          description: Null until the round locks.
        closePrice:
          oneOf:
            - $ref: "#/components/schemas/Uint256"
            - type: "null"
          description: Null until the round resolves.
        winner:
          type: string
          enum: [up, down]
          description: Present only when resolved.
        rakeTaken:
          oneOf:
            - $ref: "#/components/schemas/Uint256"
            - type: "null"
          description: Taken at resolution. Always null on a void.
        voidReason:
          type: string
          description: |
            Why the round voided, as the contract reported it — for example
            "lock window missed". Present only when voided.
          examples: ["lock window missed"]
        lastBlock:
          type: integer
          description: Block of the most recent event for this round.
    Position:
      type: object
      properties:
        round: { $ref: "#/components/schemas/Round" }
        upStake: { $ref: "#/components/schemas/Uint256" }
        downStake: { $ref: "#/components/schemas/Uint256" }
        totalStake: { $ref: "#/components/schemas/Uint256" }
        claimable:
          allOf: [{ $ref: "#/components/schemas/Uint256" }]
          description: |
            Payable now. On a void this is the stake back; on a resolved round,
            the share of the pot. Zero while the round is still running.
        claimed: { type: boolean }
        claimedAmount: { $ref: "#/components/schemas/Uint256" }
        leveraged:
          type: boolean
          description: |
            Opened with borrowed funds through the router, so settlement routes
            back through it to repay the debt rather than paying the user.
        openedAt: { type: string, format: date-time }
...